Поправил домашнее задание. Убрал лишние файлы. Объединил всё в один файл.
