--- lesson1/hw_1_2.java
@@ -1,15 +1,0 @@
-package lesson1;
-
-public class hw_1_2 {
-    public static void main(String[] args) {
-        byte a = 127;
-        short b = 1000;
-        int c = 50000;
-        long d = 10000000L;
-
-        float e = 1.5f;
-        double f = 2.5;
-
-        char g = 'g';
-    }
-}
