--- lesson1/hw_6.java
@@ -1,11 +1,0 @@
-package lesson1;
-
-public class hw_6 {
-    public static void main(String[] args) {
-        System.out.println(Calc(-5));
-    }
-    static boolean Calc (int a){
-        boolean result = a<0;
-        return result;
-    }
-}
