--- lesson1/hw_7.java
@@ -1,10 +1,0 @@
-package lesson1;
-
-public class hw_7 {
-    public static void main(String[] args) {
-        Text("John");
-    }
-    static void Text (String name){
-        System.out.println("Привет, "+name);
-    }
-}
